# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: XAUUSD Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mistakes/monthly:
    get:
      operationId: getMistakesMonthly
      tags: [trades]
      summary: Monthly breakdown of loss reasons (Mistake Journal)
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Mistake breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MistakeReport"

  # ─── HEATMAP ──────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [settings]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CONTACT ──────────────────────────────────────────────────────────────
...
  /admin/offers:
    get:
      operationId: adminListOffers
      tags: [admin]
      summary: Admin – list all offers
      responses:
        "200":
          description: All offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: adminCreateOffer
      tags: [admin]
      summary: Admin – create a new offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /admin/offers/{id}:
    patch:
      operationId: adminUpdateOffer
      tags: [admin]
      summary: Admin – update offer fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteOffer
      tags: [admin]
      summary: Admin – delete an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /admin/offers/{id}/toggle:
    patch:
      operationId: adminToggleOffer
      tags: [admin]
      summary: Admin – toggle offer active state (isOn)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Not found

  # ─── SUPPORT ──────────────────────────────────────────────────────────────
  /support/message:
    post:
      operationId: submitSupportMessage
      tags: [settings]
      summary: Authenticated user – send a support message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportMessageInput"
      responses:
        "201":
          description: Message saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportMessageResponse"
...
  /admin/support/messages/{id}/read:
    put:
      operationId: adminMarkSupportMessageRead
      tags: [admin]
      summary: Admin – mark a support message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid id

# ─── SCHEMAS ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    MistakeReasonCount:
      type: object
      required: [reason, count, percentage]
      properties:
        reason:
          type: string
        count:
          type: integer
        percentage:
          type: number
    MistakeReport:
      type: object
      required: [month, year, totalLosses, taggedLosses, reasons]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalLosses:
          type: integer
        taggedLosses:
          type: integer
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/MistakeReasonCount"

    # ACHIEVEMENTS
...
    Offer:
      type: object
      required: [id, title, isOn, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
          description: Short badge label e.g. "HOT", "LIMITED"
        price:
          type: ["number", "null"]
        originalPrice:
          type: ["number", "null"]
        isOn:
          type: boolean
          description: Whether the offer is currently active/visible to users
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        originalPrice:
          type: ["number", "null"]
        isOn:
          type: boolean

    # SUPPORT
    SupportMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    SupportMessageResponse:
      type: object
      required: [id, message, createdAt]
      properties:
        id:
          type: integer
        message:
          type: string
        createdAt:
          type: string
...